# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Exam Project API",
    "description": "Documentation for the Exam Project API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/admin/brands": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Gets brands page for admin.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/auth/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Gets all users from the database.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Login with either an email or a username, and a password. Returns a webtoken. Use this for access to other admin endpoints in this format: \"Bearer: \" + token. If the database is not filled yet, please POST \"/init\" under \"Utilities\" before proceeding.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Registers a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserRegister"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/{userId}": {
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Updates a specific user.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserEdit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "Deletes a specific user.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userName": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "address": {
                  "example": "any"
                },
                "telephoneNumber": {
                  "example": "any"
                },
                "roleId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "description": "Gets all brands from the database.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "description": "Creates a new brand.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Brand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/brands/{brandId}": {
      "put": {
        "tags": [
          "Brands"
        ],
        "description": "Updates a specific brand.",
        "parameters": [
          {
            "name": "brandId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Brand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Brands"
        ],
        "description": "Deletes a specific brand.",
        "parameters": [
          {
            "name": "brandId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "brand": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/init/": {
      "post": {
        "tags": [
          "Utilities"
        ],
        "description": "Populates the database.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "req": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "address": {
                  "example": "any"
                },
                "telephoneNumber": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Brand": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string",
          "example": "Oculus VR"
        }
      },
      "required": [
        "brand"
      ]
    },
...
    "UserLogin": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string",
          "example": "Admin"
        },
        "password": {
          "type": "string",
          "example": "P@ssword2023"
        }
      },
      "required": [
        "userName",
        "password"
      ]
    },
    "UserRegister": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "User"
        },
        "lastName": {
          "type": "string",
          "example": "NumberOne"
        },
        "userName": {
          "type": "string",
          "example": "User123"
        },
        "email": {
          "type": "string",
          "example": "user123@noroff.no"
        },
        "password": {
          "type": "string",
          "example": "P@ssword123"
        },
        "address": {
          "type": "string",
          "example": "School"
        },
        "telephoneNumber": {
          "type": "string",
          "example": "12345678"
        }
      },
      "required": [
        "firstName",
        "lastName",
        "userName",
        "email",
        "password",
        "address",
        "telephoneNumber"
      ]
    },
    "UserEdit": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "User"
        },
        "lastName": {
          "type": "string",
          "example": "NumberTwo"
        },
        "email": {
          "type": "string",
          "example": "user@noroff.no"
        },
        "address": {
          "type": "string",
          "example": "School 123"
        },
        "telephoneNumber": {
          "type": "string",
          "example": "12345678"
        },
        "roleId": {
          "type": "string",
          "example": "1"
        }
      },
      "required": [
        "firstName",
        "lastName",
        "email",
        "address",
        "telephoneNumber",
        "roleId"
      ]
    }
  }
}